// src/types/Page.ts
export type Background = Custom | Standard | Swoosh | Container;
// ...
export type Video = Vidyard | Kaltura | Html5 | Youtube;
// ...
export interface Image {
  kind?: "image";
  title?: string | null;
  src: string;
  width: number | string;
  height: number | string;
  alt?: string;
  hint?: string;
  is_decorative?: boolean;
  typeOptions?: {
    [k: string]: unknown;
  };
  sizes?: Size[];
  image_background_circle_color?: string;
  image_background_gradient_color?: string;
  image_alignment?: string;
  file_url?: string;
}
// ...
export interface Cta {
  /**
   * The label of the button.
   */
  label?: string;
  ariaLabel?: string;
  url?: string;
  /**
   * Icon identifier
   */
  icon?: string;
  variant?: "primary" | "secondary" | "tertiary" | "text" | "trial";
  target?: "_self" | "_parent" | "_top" | "_blank";
  id?: string | number;
  source?: "standard-link" | "modal-link" | "custom-action-link";
  custom_action_options?: {
    event_code?: string;
    workflow_uri?: string;
    name_capture_campaign_id?: string;
    action_template?: "everybody-demo-org" | "event-sign-up" | "offer";
    [k: string]: unknown;
  };
  linkAttributes?: {
    key?: string;
    value?: string;
    [k: string]: unknown;
  }[];
  modal_variant?: string;
  [k: string]: unknown;
}
// ...
export interface Eyebrow {
  label?: string;
  image?: Image;
  icon?: Icon;
  info_bubble?: string;
  hint?: string;
  character_count?: string;
  [k: string]: unknown;
}
export interface Icon {
  kind?: "icon";
  label?: string;
  src: string;
  decorative?: boolean;
  ariaLabel?: string;
  hint?: string;
}
export interface Custom {
  kind: "custom_background";
  blade_background_color?: string;
  blade_custom_image_desktop: Image;
  blade_custom_image_mobile: Image;
  blade_background_gradient_direction?: string;
}
export interface Standard {
  kind: "standard_background";
  blade_background_color?: string;
  blade_background_image?: Image;
  blade_background_shape?: string;
  blade_background_gradient_direction?: string;
}
export interface Swoosh {
  kind: "swoosh_background";
  blade_background_color: string;
}
export interface Container {
  kind: "container_background";
  blade_background_color?: string;
  blade_background_image?: Image;
  blade_background_image_position?: string;
  blade_background_shape?: string;
  blade_background_gradient_direction?: string;
}
// ...
export interface Media {
  image?: Image;
  video?: Video;
}
// ...
export interface Marquee {
  attributes?: {
    blade_layout: string;
    two_column_spacing?: string;
    audience_segment?: string[];
    image_overflow?: boolean;
    theme_variant?: string;
    minimize_blade_spacing_bottom?: boolean;
    minimize_blade_spacing?: boolean;
    experiment_id?: string;
    experiment_type?: string;
    eyebrow?: Eyebrow;
    use_eyebrow_as_headline?: boolean;
    headline: string;
    headline_font_color?: string;
    headline_size?: string;
    description?: string;
    anchor_id?: string;
    link_1?: Cta;
    link_2?: Cta;
    media?: Media;
    blade_background_type?: Background;
    esi_blade_flag?: string;
    esi_blade_variant?: string;
    esi_blade_rule?: string;
  };
  name: "sf/marquee";
  id: string;
  [k: string]: unknown;
}